requirements.txt
django>=1.8,<1.8.99
MySQL-python>=1.2.5,<1.2.99
jsonfield
django-geojson
django-leaflet
pil
beautifulsoup4
easy_thumbnails
django-cleanup
